# entity-service/openapi.yaml
openapi: 3.0.1
info:
  title: Entity Service API
  version: 1.0.0
servers:
  - url: "{server}:{port}/"
    variables:
      server:
        default: http://localhost
      port:
        default: "8080"
paths:
...
  /projects/{id}:
    get:
      summary: Get a project by ID.
      operationId: getProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The project with the given ID.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetailsView'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      summary: Update a project's closure-state fields or hasAgent/hasKbReferences toggles.
      description: >
        ServiceNow data source only; there is no Postgres equivalent. At least one
        field must be provided. closureState itself is not settable — it is derived
        automatically from the three closure sub-state fields by an SN business
        rule, so setting one of those and re-fetching the project is how a caller
        observes the resulting overall status.
      operationId: updateProject
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateRequest'
      responses:
        "200":
          description: The project was updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectUpdateResponse'
        "400":
          description: Bad request — no field provided, or an invalid value.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Missing or invalid x-user-id-token header (ServiceNow data source only).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/search:
    post:
      summary: Search projects by name, key, or subscription type.
      operationId: searchProjects
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchProjectsRequest'
      responses:
        "200":
          description: Projects matching the search query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchProjectsResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{id}/contacts/{contactId}:
    get:
      summary: Get one contact's attributes for a single project.
      description: >
        Returns the contact's roles on this project, their registration state and their
        notification preference. ServiceNow data source only. Returns 404 when the contact is
        not a contact on this project — including when their row has no linked contact record
        and therefore no id to match.
      operationId: getProjectContact
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: contactId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The contact's attributes for this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectContact'
        '400':
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Contact not found on this project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /projects/{id}/contacts/search:
    post:
      summary: Search contacts for a project.
      description: ServiceNow data source only; there is no Postgres equivalent.
      operationId: searchProjectContacts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchProjectContactsRequest'
      responses:
        "200":
          description: Contacts matching the search query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchProjectContactsResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Project not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /slas/{id}:
    get:
      summary: Get a task SLA record by ID (ServiceNow data source only).
      operationId: getTaskSla
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The task SLA record.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSlaDetail'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "404":
          description: Task SLA not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /slas/search:
    post:
      summary: Search task SLA records (ServiceNow data source only).
      operationId: searchTaskSlas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchTaskSlasRequest'
      responses:
        "200":
          description: Task SLA records matching the filters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchTaskSlasResponse'
        "400":
          description: Bad request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          description: Number of results per page. Defaults to 20, capped at 50.
        offset:
          type: integer
          description: Zero-based offset into the result set.
...
    SearchProjectsRequest:
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/Pagination'
        searchQuery:
          type: string
          description: Case-insensitive match against name, key, and subscription_type.
    ProjectAccountRef:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        activationDate:
          type: string
          format: date-time
          nullable: true
        tier:
          type: string
        region:
          type: string
          nullable: true
        agentEnabled:
          type: boolean
        kbReferencesEnabled:
          type: boolean
    ProjectDetailsView:
      type: object
      properties:
        id:
          type: string
        account:
          $ref: '#/components/schemas/ProjectAccountRef'
        sfId:
          type: string
        name:
          type: string
        key:
          type: string
        subscriptionType:
          type: string
          enum:
            - development_support
            - managed_cloud_subscription
            - evaluation_subscription
            - subscription
            - cloud_evaluation_support
            - internal
            - platformer_subscription
            - cloud_support
            - professional_services
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        createdOn:
          type: string
          format: date-time
        updatedOn:
          type: string
          format: date-time
        closureState:
          type: string
          nullable: true
        endDateClosureState:
          type: string
          nullable: true
        invoiceDueDateClosureState:
          type: string
          nullable: true
        complianceViolationClosureState:
          type: string
          nullable: true
        complianceViolationDate:
          type: string
          nullable: true
        suspensionProcessState:
          type: object
          additionalProperties: true
          nullable: true
          description: >
            Free-form, per-dimension Account Closure Process (ACP) suspension-process
            state (event type + action results) written by an existing ServiceNow
            suspension flow. This is opaque JSON passed through as-is — its shape is
            not defined or validated by this layer.
    ProjectView:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
        subscriptionType:
          type: string
          enum:
            - development_support
            - managed_cloud_subscription
            - evaluation_subscription
            - subscription
            - cloud_evaluation_support
            - internal
            - platformer_subscription
            - cloud_support
            - professional_services
        startDate:
          type: string
          format: date-time
          nullable: true
          description: >
            Start of the project's current renewed period. Null when the backing
            data source has no start date recorded.
        endDate:
          type: string
          format: date-time
          nullable: true
          description: End of the project's current renewed period.
        createdOn:
          type: string
          format: date-time
          description: >
            When the project record was created. Distinct from startDate, which
            moves forward on each renewal.
        account:
          nullable: true
          description: Null when the project has no linked account.
          allOf:
            - $ref: '#/components/schemas/EntityRef'
        closureState:
          type: string
          nullable: true
        endDateClosureState:
          type: string
          nullable: true
        invoiceDueDateClosureState:
          type: string
          nullable: true
        complianceViolationClosureState:
          type: string
          nullable: true
        complianceViolationDate:
          type: string
          nullable: true
        suspensionProcessState:
          type: object
          additionalProperties: true
          nullable: true
          description: >
            Free-form, per-dimension Account Closure Process (ACP) suspension-process
            state (event type + action results) written by an existing ServiceNow
            suspension flow. This is opaque JSON passed through as-is — its shape is
            not defined or validated by this layer.
    SearchProjectsResponse:
      type: object
      properties:
        projects:
          type: array
          items:
            $ref: '#/components/schemas/ProjectView'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        hasMore:
          type: boolean
    ProjectUpdateRequest:
      type: object
      minProperties: 1
      additionalProperties: false
      description: >
        At least one field must be provided. hasAgent and hasKbReferences are
        simple toggles; the closure-state fields track sub-states of an
        Account Closure Process and are only applicable to projects sourced
        from ServiceNow.
      properties:
        hasAgent:
          type: boolean
        hasKbReferences:
          type: boolean
        endDateClosureState:
          type: string
        invoiceDueDateClosureState:
          type: string
        complianceViolationClosureState:
          type: string
        suspensionProcessState:
          type: object
          additionalProperties: true
          nullable: true
          description: >
            Free-form, per-dimension Account Closure Process (ACP) suspension-process
            state (event type + action results) written by an existing ServiceNow
            suspension flow. This is opaque JSON passed through as-is — its shape is
            not defined or validated by this layer.
    ProjectUpdateResponse:
      type: object
      required: [message, project]
      properties:
        message:
          type: string
        project:
          $ref: '#/components/schemas/ProjectUpdateResult'
    ProjectUpdateResult:
      type: object
      required: [id, updatedBy, updatedOn]
      properties:
        id:
          type: string
        updatedBy:
          type: string
        updatedOn:
          type: string
          format: date-time
        closureState:
          type: string
          nullable: true
        endDateClosureState:
          type: string
          nullable: true
        invoiceDueDateClosureState:
          type: string
          nullable: true
        complianceViolationClosureState:
          type: string
          nullable: true
        suspensionProcessState:
          type: object
          additionalProperties: true
          nullable: true
    SearchProjectContactsRequest:
      type: object
      properties:
        filters:
          type: object
          properties:
            searchQuery:
              type: string
              description: Searches across contact name and email (case-insensitive).
        pagination:
          $ref: '#/components/schemas/Pagination'
    ProjectContact:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: >
            The contact's user id, for linking the row to that user's profile. Absent when the
            row has no contact record linked, or when the backing instance predates the field —
            render the row unlinked rather than treating it as an error.
        name:
          type: string
        email:
          type: string
        registrationState:
          type: string
        notificationsEnabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
        customerContactPresent:
          type: boolean
          description: >
            Whether a contact record is linked to this row at all. False is the same fault an
            absent `id` signals, restated as an explicit boolean.
        grantsCaseAccess:
          type: boolean
          description: >
            Whether this row would actually grant its person visibility into this project's
            cases. Mirrors `customerContactPresent` directly — a row can be listed here without
            granting access.
    SearchProjectContactsResponse:
      type: object
      properties:
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/ProjectContact'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
...
    EntityRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
    SearchTaskSlasRequest:
      type: object
      properties:
        filters:
          type: object
          properties:
            taskIds:
              type: array
              items:
                type: string
                format: uuid
        pagination:
          type: object
          properties:
            offset:
              type: integer
              minimum: 0
              default: 0
            limit:
              type: integer
              minimum: 1
              maximum: 50
              default: 20
    TaskSlaDefinition:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        target:
          type: string
          nullable: true
    TaskSlaTaskRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
    TaskSlaView:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
        slaDefinition:
          $ref: '#/components/schemas/TaskSlaDefinition'
          nullable: true
        stage:
          type: string
          nullable: true
        task:
          $ref: '#/components/schemas/TaskSlaTaskRef'
          nullable: true
        businessTimeLeft:
          type: string
          nullable: true
        businessElapsedTime:
          type: string
          nullable: true
        businessElapsedPercentage:
          type: number
          format: double
          nullable: true
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
    SearchTaskSlasResponse:
      type: object
      properties:
        slas:
          type: array
          items:
            $ref: '#/components/schemas/TaskSlaView'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    TaskSlaDefinitionDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          nullable: true
        type:
          type: string
          nullable: true
        target:
          type: string
          nullable: true
        flow:
          type: string
          nullable: true
        workflow:
          type: string
          nullable: true
        isEnableLogging:
          type: boolean
          nullable: true
        durationType:
          type: string
          nullable: true
        duration:
          type: string
          nullable: true
        scheduleSource:
          type: string
          nullable: true
        schedule:
          type: string
          nullable: true
        timezoneSource:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
        startCondition:
          type: string
          nullable: true
        isRetroactiveStart:
          type: boolean
          nullable: true
        isRetroactivePause:
          type: boolean
          nullable: true
        whenToCancel:
          type: string
          nullable: true
        cancelCondition:
          type: string
          nullable: true
        pauseCondition:
          type: string
          nullable: true
        whenToResume:
          type: string
          nullable: true
        stopCondition:
          type: string
          nullable: true
        resetCondition:
          type: string
          nullable: true
        resetAction:
          type: string
          nullable: true
    TaskSlaScheduleRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
          nullable: true
        name:
          type: string
          nullable: true
        timezone:
          type: string
          nullable: true
    TaskSlaDetail:
      type: object
      required: [id]
      properties:
        id:
          type: string
          format: uuid
        task:
          $ref: '#/components/schemas/TaskSlaTaskRef'
          nullable: true
        slaDefinition:
          $ref: '#/components/schemas/TaskSlaDefinitionDetail'
          nullable: true
        stage:
          type: string
          nullable: true
        businessTimeLeft:
          type: string
          nullable: true
        businessElapsedTime:
          type: string
          nullable: true
        businessElapsedPercentage:
          type: number
          format: double
          nullable: true
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        active:
          type: boolean
          nullable: true
        schedule:
          $ref: '#/components/schemas/TaskSlaScheduleRef'
          nullable: true
...